431: Remove toastContext. Add showToast hook using rtk store
