--- packages/orchestrator-ui-components/src/hooks/useToastMessage.ts
@@ -1,5 +1,0 @@
-import { useContext } from 'react';
-
-import { ToastContext } from '../contexts/ToastContext';
-
-export const useToastMessage = () => useContext(ToastContext);
